refactor(api): migrate certification-point-of-contact repository
in identity-access-management
